refactor: extract buildInlineDiffSegments to shared diff-utils module
Deduplicate buildInlineDiffSegments and diffCharsIfWithinEditLimit from
src/api/diff.ts, src/browser/lib/diff-worker.ts, and
src/browser/ui/diff/utils/parse.ts into src/shared/diff-utils.ts.

Callers now pass (currentContent, nextContent, inlineMaxCharEdits) as
plain values instead of (_Change, _Change, ParseOptions).

Closes #20
